Implement replay fallback policy for subscription failures
If the Client is opened using a ReplayMarkerStorage
implementation which haven't been updated in the last 24 hours,
then on subscription, Salesforce will get a replay id for a
message it no longer has. In these cases a ServerError would be
raised, and users would have to handle it themselves. To ease the
handling of these kind of problems, define a fallback replay
policy. On ServerError the Client will retry to subscribe using
the replay id specified in the replay_fallback parameter.