Extract replay id from incoming messages in Client.receive
Wait for the extraction of replay ids from incoming messages until
we become sure that they're going to be consumed. If the replay id
is extracted in the ReplayMarkerStorage extension when the
message is received, then if the Client disconnects without
consuming all prefetched messages, then the next time it
reconnects, it will miss the unconsumed and prefetched messages
from the previous run.