Store replay ids which have the same date as the previous one
For mass record delete operations on Salesforce, all the delete
events will have the same date. So when a replay id in an incoming
message is compared to the previous replay id, the new id should
be stored not just if it is newer, but also when it has exactly
the same date.