Implement Client.__aiter__ as an asynchronous generator
According to PEP 525 returning an awaitable resulting in an
asynchronous generator has been deprecated. So __aiter__ should
be either a regular function returning an asynchronous iterator
or it should be an asynchronous generator function. Implement
the __aiter__ method of Client as an asynchronous generator
yielding the values yielded from the __aiter__ method of the
super class.